Remove date label
This label was supposed to help identify the cost with the date it
represents. But the label would have unbounded values and high
cardinality. This is not recommended in prometheus:
https://prometheus.io/docs/practices/naming/#labels

The metric have an associated timestamp anyway. We have to remember that
a timestamp identifies the cost of the PREVIOUS DAY.