Adjust postgres storage price
As reported in ticket 4665757 with GDS, the API returns the wrong price
for postgres storage price. It returns a fixed monthly price, whatever
the chosen time range.
As a workaround, we adjust the price ourselves.

This is only temporary as we're expecting a new billing API which won't
have this issue.